+++ extras/tests/TestCodeRealDataPackageMode.R
@@ -39,9 +39,10 @@ cohortTable <- # example: 'cohort'
 
 outputFolder <-
   file.path(rstudioapi::getActiveProject(), "outputFolder", databaseId)
-unlink(x = outputFolder,
-       recursive = TRUE,
-       force = TRUE)
+# Please delete previous content if needed
+# unlink(x = outputFolder,
+#        recursive = TRUE,
+#        force = TRUE)
 dir.create(path = outputFolder,
            showWarnings = FALSE,
            recursive = TRUE)
@@ -63,7 +64,7 @@ SkeletonCohortDiagnosticsStudy::runCohortDiagnostics(
   databaseId = databaseId,
   databaseName = dataSouceInformation$cdmSourceName,
   databaseDescription = dataSouceInformation$sourceDescription,
-  runCohortCharacterization = TRUE,
+  runCohortCharacterization = FALSE,
   runCohortOverlap = TRUE,
   runOrphanConcepts = TRUE,
   runVisitContext = TRUE,