+++ tests/testthat/test-againstEunomia.R
@@ -30,8 +30,8 @@ test_that("Cohort instantiation", {
   
   connection <- DatabaseConnector::connect(connectionDetails)
   sql <-
-    "SELECT COUNT(*) AS cohort_count, cohort_definition_id 
-  FROM @cohort_database_schema.@cohort_table 
+    "SELECT COUNT(*) AS cohort_count, cohort_definition_id
+  FROM @cohort_database_schema.@cohort_table
   GROUP BY cohort_definition_id;"
   counts <-
     DatabaseConnector::renderTranslateQuerySql(
@@ -105,9 +105,7 @@ test_that("Cohort diagnostics in incremental mode", {
   
   # generate premerged file
   CohortDiagnostics::preMergeDiagnosticsFiles(dataFolder = file.path(folder, "export"))
-  testthat::expect_true(file.exists(file.path(
-    folder, "export", "PreMerged.RData"
-  )))
+  testthat::expect_true(file.exists(file.path(folder, "export", "PreMerged.RData")))
 })
 
 unlink(folder, recursive = TRUE)