+++ NEWS.md
@@ -18,6 +18,7 @@ Changes:
 7. Run against external concept count has been removed, as concept counts data is not available. Function 'runCohortDiagnosticsUsingExternalCounts' is removed.
 8. Removed code related to referentConceptId = phenotypeId/1000 as it does not always hold true.
 9. Create cohort table function is now private. Please use instantiate cohort.
+10. checkInputFileEncoding is not exported as a public function (as not the scope of CohortDiagnostics).
 
 Bug fixes: